requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.9
oauthlib==3.1.0
praw==7.0.0
prawcore==1.4.0
PySocks==1.7.1
python-dotenv==0.13.0
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.8.0
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
